requirements.txt
azure-identity==1.19.0
azure-storage-blob==12.24.0
python-dotenv==1.0.1
